requirements.txt
fastapi==0.111.1
uvicorn==0.25.0
SQLAlchemy==2.0.20
pydantic==2.6.2
python-jose==3.3.0
argon2-cffi==23.1.0